Modified navigation drawer to include today, all and high prioirity items.
