+++ app/src/main/res/menu/menu_fragment_all.xml
