+++ app/src/main/res/values/strings.xml
@@ -17,7 +17,7 @@
     <string name="action_delete">Delete</string>
     <string name="action_delete_all">Delete All</string>
 
-    <string name="drawer_label_home">Home</string>
+    <string name="drawer_label_home">Today Items</string>
     <string name="drawer_label_all">All Items</string>
     <string name="drawer_label_high">High Priority Items</string>
     <string name="drawer_open">Open navigation drawer</string>
